Make SQLiteOnlyExpandedDatabase support preserving ordering.
We need to use OrderedDict to preserve the ordering when converting
into/out-of JSON. This was introduced in py2.7, but there is a compatible
py2.6 implementation we can use.
We also need an extra column in the document_fields table, to indicate
what order the field comes so that it stays preserved.

We now implement the full base api.