It appears that pkg_resources is a reasonable way to get at the data file.
I'm not sure if it is ideal to depend on pkg_resources, since setup.py has code to
fallback to distutils.core instead of strictly depending on setuptools being available.
But it would mean that we can get the data content from a .zip file if it all gets
packaged up on Windows, etc.