Use the common dbschema as part of the C code.
This uses a python script to just change an SQL file into a C array which we then iterate over.
We had one discrepancy which was 'db_rev' vs 'generation' in the SQL definitions.

This isn't the prettiest parser or anything, but it works, and we can always extend it as we
need fancier SQL. (for example, it doesn't escape quote characters.)

I wanted to build it from setup.py, but that looks hard to do, so I just do
it from the Makefile.