Cleanup the u1db_delete_doc code.
This makes sure that it uses the right logic to create a new revision after a delete.
We also had some tests that assumed delete_doc would set the content to None, so add
an explicit test for that behavior (and implement it in the C code).


Also, clean up the code a bit using the goto structure. The rest of the code uses it
and it makes the code quite a bit cleaner.