Get rid of having entries added to a query object.
Instead, just have get_from_index handle 1 item at a time.
There are bits of caching that might be possible if we find we need it,
namely turning a specific number of query parameters into an actual query.
Or possibly even caching the prepared sqlite statement object.
However, the interface of multiple requests is really clumsy, so lets punt
for now. If we really want, we could do things like caching stuff on the
u1query object.

For now, we keep the u1query object, so you don't have to do the db lookup
of the index definition for each request.