GCC doesn't like our static array construction.
Specifically, it forces 'const char *var[]' to be an array of size 1,
rather than currently undefined. So instead we create it as a:
'const char **var', and then use a local static tmp variable that
holds the 'const char *tmp[] = {...}'. This makes the compiler
happy, and the tests still pass.