Finish implementing simple wildcard searching.
Simplify the main bind loop. We don't re-use the structure at all, so we just bind the
arguments in order, and it makes a lot more sense to the code.

Change format_query since now as it determines what bits need to be in the query, it
also determines whether a given request is a wildcard or not.