Start looking at the find_changes_to_return api.
It turns out that we do need the 'find' step first, because we need the 'new_gen' that
we get from whats_changed in order to set the header for the HTTP api.
I'll see if I can do something about that in the C code. For now, we don't need to
pass more arguments to the SyncExchange object, rather than just passing them in
the constructor.