Move the conflict tests from sync into backend tests.
We can use put_doc_if_newer to introduce conflicts, we don't have to do it
via sync, which means it is a more direct test, and can be easily written
against the C backend.

Fix u1db_put_doc so it checks for conflicts, still need to do delete doc, etc.