A fair number of changes.
We have to declare the C api functions as 'nogil' because we need the pure-python HTTP thread
to be able to run while the curl code is in perform() waiting for the response.

Lots of code to set up an http request, and then read the response from the server.
Then parse it into json, and read out the content.
Still need to do a lot more refactoring to get this into re-usable code, but at least
something is actually talking HTTP to someone else.