Implement record_sync_info.
I need a bit more time to work out how to share the CURL object correctly, especially
how to handle a HTTP/1.0 vs HTTP/1.1 connection.
The test server is a 1.0 server, and doesn't support 100-continue on PUT. Which we want
to have CURL remember.
We may have to manually tell curl if we encounter a 1.0 server, even though it should
be noticing that itself. Maybe I'm doing something wrong.
Also, we may/may not want to be calling curl_easy_reset before or after each request.
It seems to reset a bit more than we really want to, but as I said, I need to play with
it a bit more.