getting part of the http sync exchange working.
Still a bit tricky, Windows doesn't have mkstemp, so we brought in an implementation.
Curl doesn't want to let us just HTTP POST because it only wants us to POST by
supplying the form data. I need to figure out how to work around that.

So not quite working, but getting closer.