Fix a memory leak in the http_get_sync_info code.
The api for get_sync_info says it returns a const char *, which means
the caller should not be freeing the memory. This is because the
direct C impl just returns the cached value. So we cache the value
in the HTTP implementation because it is safe to do so. (If the
replica_uid would actually change, things would be really broken.)