Free documents during sync_exchange
During sync_exchange we call get_and_insert_docs which calls
get_docs and passes the callback to insert for synchronization.
However, we weren't freeing those documents because insert shouldn't
be freeing them, but sync code doesn't see them.
So update the get_docs_to_doc_gen code to allow a flag to indicate
the callback isn't going to free the docs.
This is about 1.6kB during the test suite.