Squash the new memory leak.
We were passing a parameter (that isn't in use yet), which
was then getting overwritten with the return value.
Document that we will be doing this (the parameter is IN/OUT
and the caller is responsible for freeing *both* of them).
And then fix the sync code to use another variable so
that it can be freed appropriately.